refactor: context, hooks, utils fn and types move
