+++ types/peopleType.ts
