+++ utils/createPageNums.test.ts
@@ -1,4 +1,4 @@
-import { createPageNums } from '../../utils/createPageNums';
+import { createPageNums } from './createPageNums';
 
 describe('createPageNums', () => {
   it('should return an array with one page number for totalItems less than 10', () => {
