+++ utils/findSelectedPerson.ts
